+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `create-config` would fail with `--path` specified.
+- `migrate` no longer fails if the `cache` directory does not exist.
 
 ## [2.1.1] - 2022-05-29